# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deck Me API — decking quotes, materials, customers, and bookings.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      parameters:
        - in: query
          name: category
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Material" }
    post:
      operationId: createMaterial
      tags: [materials]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Material" }
  /materials/{id}:
    patch:
      operationId: updateMaterial
      tags: [materials]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MaterialUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Material" }
    delete:
      operationId: deleteMaterial
      tags: [materials]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletionSuccess" }
...
  /master-projects:
    get:
      operationId: listMasterProjects
      tags: [master-projects]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MasterProjectSummary" }
    post:
      operationId: createMasterProject
      tags: [master-projects]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterProjectInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MasterProject" }
  /master-projects/{id}:
    get:
      operationId: getMasterProject
      tags: [master-projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MasterProject" }
    patch:
      operationId: updateMasterProject
      tags: [master-projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterProjectUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MasterProject" }
    delete:
      operationId: deleteMasterProject
      tags: [master-projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuccessWithId" }
  /master-projects/{id}/quotes:
    put:
      operationId: setMasterProjectQuotes
      tags: [master-projects]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterProjectQuoteSelection" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MasterProject" }
...
components:
  schemas:
...
    DeletionSuccess:
      type: object
      required: [deleted]
      properties:
        deleted: { type: boolean, const: true }
...
    SuccessWithId:
      type: object
      required: [success, id]
      properties:
        success: { type: boolean, const: true }
        id: { type: integer }
...
    Material:
      type: object
      required: [id, name, category, unit, unitPrice, supplier]
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: ["string", "null"] }
        category:
          type: string
          description: "decking | joist | bearer | post | stump | screw | nail | bracket | sealant | other"
        unit:
          type: string
          description: "each | metre | pack | bag"
        unitPrice: { type: number, description: "Retail price per unit (AUD) — shown on client quotes" }
        tradeCost: { type: ["number", "null"], description: "Your trade/buy price — never shown to clients" }
        packSize: { type: ["integer", "null"], description: "Items per pack, if unit=pack" }
        supplier:
          type: string
          description: "bunnings | mitre10 | local | other"
        notes: { type: ["string", "null"] }
    MaterialInput:
      type: object
      required: [name, category, unit, unitPrice, supplier]
      properties:
        name: { type: string, minLength: 1 }
        sku: { type: string }
        category: { type: string }
        unit: { type: string }
        unitPrice: { type: number }
        tradeCost: { type: number }
        packSize: { type: integer }
        supplier: { type: string }
        notes: { type: string }
    MaterialUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        sku: { type: string }
        category: { type: string }
        unit: { type: string }
        unitPrice: { type: number }
        tradeCost: { type: number }
        packSize: { type: integer }
        supplier: { type: string }
        notes: { type: string }
...
    MasterProjectSummary:
      type: object
      required: [id, title, status, customerId, builderMarginPct, materialsSubtotal, labourSubtotal, marginAmount, gst, total, quoteCount, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        status: { type: string }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        builderMarginPct: { type: number }
        materialsSubtotal: { type: number }
        labourSubtotal: { type: number }
        marginAmount: { type: number }
        gst: { type: number }
        total: { type: number }
        quoteCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MasterProjectQuote:
      type: object
      required: [id, title, status, customerId, tradeType, materialsSubtotal, labourCost, gst, total, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        status: { type: string }
        customerId: { type: integer }
        tradeType: { type: string }
        materialsSubtotal: { type: number }
        labourCost: { type: number }
        gst: { type: number }
        total: { type: number }
        createdAt: { type: string, format: date-time }
    MasterBillOfMaterialsLine:
      type: object
      required: [description, category, quantity, unit, unitPrice, lineTotal, quoteCount]
      properties:
        materialId: { type: ["integer", "null"] }
        description: { type: string }
        category: { type: string }
        quantity: { type: number }
        unit: { type: string }
        unitPrice: { type: number }
        lineTotal: { type: number }
        quoteCount: { type: integer }
    MasterProject:
      type: object
      required: [id, title, status, customerId, builderMarginPct, materialsSubtotal, labourSubtotal, marginAmount, gst, total, quotes, billOfMaterials, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        status: { type: string }
        customerId: { type: integer }
        customerName: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        builderMarginPct: { type: number }
        materialsSubtotal: { type: number }
        labourSubtotal: { type: number }
        marginAmount: { type: number }
        gst: { type: number }
        total: { type: number }
        quotes:
          type: array
          items: { $ref: "#/components/schemas/MasterProjectQuote" }
        billOfMaterials:
          type: array
          items: { $ref: "#/components/schemas/MasterBillOfMaterialsLine" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MasterProjectInput:
      type: object
      required: [title, customerId]
      properties:
        title: { type: string, minLength: 1 }
        customerId: { type: integer }
        builderMarginPct: { type: number, minimum: 0, maximum: 99.99, default: 0 }
        notes: { type: string }
    MasterProjectUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        status: { type: string }
        builderMarginPct: { type: number, minimum: 0, maximum: 99.99 }
        notes: { type: ["string", "null"] }
    MasterProjectQuoteSelection:
      type: object
      required: [quoteIds]
      properties:
        quoteIds:
          type: array
          uniqueItems: true
          items: { type: integer }
...