# lib/api-zod/src/generated/types/successWithId.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deck Me API — decking quotes, materials, customers, and bookings.
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessWithId {
  success: true;
  id: number;
}